Modification de la base de données

+++ src/main/java/devops/kilroywashere/wshabitation/repositories/HabitationsRepository.java
@@ -3,5 +3,5 @@ package devops.kilroywashere.wshabitation.repositories;
 import devops.kilroywashere.wshabitation.models.Habitation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface HabitationsRepository extends CrudRepository<Habitation, Long> {
+public interface HabitationsRepository extends CrudRepository<Habitation, Integer> {
 }
